+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+* v.0.4.0
+  * Introducing $translateProvider.rememberLanguage()
+  * You're now able to tell ngTranslate save lang state cross requests
 * v.0.3.0
   * $translate Service now has method uses(key) to ask for currently used language
   * Language can now be changed at runtime